+++ R/package.R
@@ -29,10 +29,10 @@
 #' @importFrom Matrix Matrix sparseMatrix sparseVector crossprod tcrossprod
 #' drop0 nnzero
 #' @importFrom Biobase assayData
-#' @importFrom ggplot2 ggplot geom_raster aes_string facet_grid
-#' scale_fill_manual scale_alpha scale_x_discrete scale_y_continuous geom_vline
-#' geom_hline ggtitle theme element_text element_blank element_rect
-#'
+#' @importFrom ggplot2 aes aes_string element_blank element_rect element_text
+#' facet_grid geom_hline geom_raster geom_segment geom_text geom_vline ggplot
+#' ggtitle labs scale_alpha scale_color_manual scale_fill_manual
+#' scale_x_discrete scale_y_continuous theme theme_classic
 #' @importFrom mzR openMSfile close header peaks runInfo
 #' @importFrom MSnbase calculateFragments defaultNeutralLoss get.amino.acids
 #' @importClassesFrom MSnbase MSnSet